pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clustkit"
version = "0.1.0"
description = "Accurate sequence clustering and search via adaptive banded alignment"
license = "MIT"
requires-python = ">=3.10"
readme = "README.md"
authors = [
    { name = "ClustKIT Contributors" },
]
keywords = [
    "bioinformatics", "clustering", "sequence-search", "alignment",
    "MinHash", "LSH", "Needleman-Wunsch", "protein", "nucleotide",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "numpy>=1.24.0",
    "numba>=0.58.0",
    "biopython>=1.82",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
leiden = [
    "leidenalg",
    "python-igraph",
]
gpu = [
    "cupy-cuda12x",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "scikit-learn>=1.3.0",
]
benchmarks = [
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
]

[project.scripts]
clustkit = "clustkit.cli:app"

[project.urls]
Homepage = "https://github.com/clustkit/clustkit"
Repository = "https://github.com/clustkit/clustkit"
Issues = "https://github.com/clustkit/clustkit/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["clustkit*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
